Move LiveRange and LivenessInfo to shared regalloc module
Consolidate liveness analysis types with register allocation in
crate::regalloc. The old crate::liveness module now re-exports
from regalloc for backward compatibility.

Fixes rue-q6uu
