+++ crates/rue-codegen/src/regalloc.rs
@@ -1,10 +1,18 @@
 //! Shared register allocation algorithm and types.
 //!
-//! This module provides a target-independent linear scan register allocator.
-//! Each backend implements the `RegAllocBackend` trait to provide target-specific
-//! details like available registers and instruction rewriting.
+//! This module provides a target-independent linear scan register allocator
+//! and the liveness analysis types used by all backends.
 //!
-//! ## Algorithm
+//! ## Liveness Analysis
+//!
+//! The module provides target-independent types for liveness analysis:
+//! - [`LiveRange`]: Represents the instruction range where a vreg's value is needed
+//! - [`LivenessInfo`]: Holds all liveness information (ranges, live_at, clobbers)
+//!
+//! Each backend implements its own `analyze()` function that populates these types
+//! based on its specific instruction set and control flow.
+//!
+//! ## Register Allocation Algorithm
 //!
 //! The allocator uses linear scan register allocation:
 //! 1. Compute live ranges for all virtual registers (via liveness analysis)
@@ -19,11 +27,121 @@
 //! the longest remaining live range, as this frees up a register for the
 //! longest time.
 
-use std::collections::HashSet;
+use std::collections::{HashMap, HashSet};
 
 use crate::index_map::IndexMap;
-use crate::liveness::{LiveRange, LivenessInfo};
 use crate::vreg::VReg;
+
+// ============================================================================
+// Liveness Analysis Types
+// ============================================================================
+
+/// Live range for a virtual register.
+///
+/// Represents the instruction range where this vreg's value is needed.
+/// Live ranges are [start, end] inclusive intervals.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct LiveRange {
+    /// Instruction index where the vreg is defined (first write).
+    pub start: usize,
+    /// Instruction index where the vreg is last used (last read).
+    pub end: usize,
+}
+
+impl LiveRange {
+    /// Create a new live range.
+    #[inline]
+    pub const fn new(start: usize, end: usize) -> Self {
+        Self { start, end }
+    }
+
+    /// Check if this live range overlaps with another.
+    ///
+    /// Two ranges overlap if they share at least one instruction index.
+    #[inline]
+    pub fn overlaps(&self, other: &LiveRange) -> bool {
+        self.start <= other.end && other.start <= self.end
+    }
+}
+
+/// Result of liveness analysis.
+///
+/// This struct is target-independent and holds all the information needed
+/// by the register allocator. Each backend's `analyze()` function populates
+/// an instance of this type.
+pub struct LivenessInfo<Reg: Copy + Eq + std::hash::Hash> {
+    /// Live range for each virtual register.
+    pub ranges: HashMap<VReg, LiveRange>,
+    /// For each instruction, which vregs are live after it executes.
+    /// This is useful for determining which registers are in use at any point.
+    pub live_at: Vec<HashSet<VReg>>,
+    /// For each instruction index, the physical registers clobbered by that instruction.
+    /// This is used to prevent allocating vregs to registers that would be clobbered.
+    pub clobbers_at: Vec<Vec<Reg>>,
+}
+
+impl<Reg: Copy + Eq + std::hash::Hash> LivenessInfo<Reg> {
+    /// Create a new empty liveness info.
+    pub fn new() -> Self {
+        Self {
+            ranges: HashMap::new(),
+            live_at: Vec::new(),
+            clobbers_at: Vec::new(),
+        }
+    }
+
+    /// Get vregs that are live at a given instruction index.
+    pub fn live_at(&self, inst_idx: usize) -> &HashSet<VReg> {
+        &self.live_at[inst_idx]
+    }
+
+    /// Get the live range for a vreg.
+    pub fn range(&self, vreg: VReg) -> Option<&LiveRange> {
+        self.ranges.get(&vreg)
+    }
+
+    /// Check if two vregs interfere (have overlapping live ranges).
+    ///
+    /// Two vregs interfere if they are both live at the same program point,
+    /// meaning they cannot share the same physical register.
+    pub fn interferes(&self, a: VReg, b: VReg) -> bool {
+        match (self.ranges.get(&a), self.ranges.get(&b)) {
+            (Some(ra), Some(rb)) => ra.overlaps(rb),
+            _ => false,
+        }
+    }
+
+    /// Get the physical registers clobbered at a given instruction index.
+    pub fn clobbers_at(&self, inst_idx: usize) -> &[Reg] {
+        &self.clobbers_at[inst_idx]
+    }
+
+    /// Check if a physical register is clobbered while a vreg is live.
+    ///
+    /// Returns true if `reg` is clobbered by any instruction during the live range of `vreg`.
+    /// This is used to prevent allocating a vreg to a register that would be clobbered
+    /// before the vreg's last use.
+    pub fn is_clobbered_during(&self, vreg: VReg, reg: Reg) -> bool {
+        if let Some(range) = self.ranges.get(&vreg) {
+            for idx in range.start..=range.end {
+                if idx < self.clobbers_at.len() && self.clobbers_at[idx].contains(&reg) {
+                    return true;
+                }
+            }
+        }
+        false
+    }
+}
+
+impl<Reg: Copy + Eq + std::hash::Hash> Default for LivenessInfo<Reg> {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+// ============================================================================
+// Register Allocation Types
+// ============================================================================
 
 /// Allocation result for a virtual register.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
@@ -152,8 +270,51 @@ pub fn linear_scan<Reg: Copy + Eq + std::hash::Hash>(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::liveness::LiveRange;
-    use std::collections::HashMap;
+
+    // ========================================
+    // LiveRange tests
+    // ========================================
+
+    #[test]
+    fn test_live_range_overlaps() {
+        let r1 = LiveRange::new(0, 5);
+        let r2 = LiveRange::new(3, 8);
+        let r3 = LiveRange::new(6, 10);
+
+        // r1 and r2 overlap at 3-5
+        assert!(r1.overlaps(&r2));
+        assert!(r2.overlaps(&r1));
+
+        // r1 and r3 don't overlap (r1 ends at 5, r3 starts at 6)
+        assert!(!r1.overlaps(&r3));
+        assert!(!r3.overlaps(&r1));
+
+        // r2 and r3 overlap at 6-8
+        assert!(r2.overlaps(&r3));
+        assert!(r3.overlaps(&r2));
+    }
+
+    #[test]
+    fn test_live_range_adjacent_not_overlapping() {
+        // Adjacent ranges should overlap (inclusive end)
+        let r1 = LiveRange::new(0, 5);
+        let r2 = LiveRange::new(5, 10);
+
+        // At instruction 5, both ranges are active
+        assert!(r1.overlaps(&r2));
+    }
+
+    #[test]
+    fn test_live_range_same_point() {
+        let r1 = LiveRange::new(5, 5);
+        let r2 = LiveRange::new(5, 5);
+
+        assert!(r1.overlaps(&r2));
+    }
+
+    // ========================================
+    // Linear scan allocation tests
+    // ========================================
 
     // Simple test register type
     #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
